change probe table view to tree view and split controller
