+++ ptychodus/controller/probe/__init__.py
@@ -1,0 +1,7 @@
+from .controller import ProbeController
+from .image import ProbeImageController
+
+__all__ = [
+    'ProbeController',
+    'ProbeImageController',
+]
